refactor: align Card presentation with FSD
