+++ src/entities/card/ui/BackText.spec.tsx
