+++ src/entities/card/ui/CardView.spec.tsx
