+++ src/entities/card/ui/CardView.tsx
