+++ src/entities/card/ui/FrontText.spec.tsx
