--- src/features/card-view/index.ts
@@ -1,4 +1,0 @@
-export { BackText } from "./ui/BackText";
-export { CardOverlay } from "./ui/CardOverlay";
-export { CardView } from "./ui/CardView";
-export { FrontText } from "./ui/FrontText";
