+++ src/pages/study-session/ui/CardOverlay.spec.tsx
